requirements.txt
pandas~=2.2.3
future~=1.0.0
matplotlib~=3.9.2
tksheet~=7.2.21
numpy~=2.1.3
seaborn~=0.13.2
scikit-learn~=1.5.2
pillow~=11.0.0
requests~=2.32.3
opencv-python~=4.10.0.84